+++ packages/components/CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Bug Fix
 
 -    The `Button` component now displays the label as the tooltip for icon only buttons. ([#40716](https://github.com/WordPress/gutenberg/pull/40716))
+-    Use fake timers and fix usage of async methods from `@testing-library/user-event`. ([#40790](https://github.com/WordPress/gutenberg/pull/40790))
 
 ### Internal